backend/requirements.txt
# API framework
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9
websockets==12.0

# LangChain + LangGraph
langchain==0.2.16
langchain-anthropic==0.1.23
langchain-community==0.2.16
langgraph==0.2.6
langsmith==0.1.98

# Async task queue
celery==5.4.0
redis==5.0.4
flower==2.0.1

# Vector DB
pinecone-client==4.1.0

# Document parsing
pypdf==4.2.0
beautifulsoup4==4.12.3
requests==2.32.2
youtube-transcript-api==0.6.2

# Database
sqlalchemy==2.0.30
alembic==1.13.1
psycopg2-binary==2.9.9

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# Graph database
neo4j==5.20.0

# Utilities
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
httpx==0.27.0
tenacity==8.3.0
apscheduler==3.10.4
sentence-transformers>=2.7.0
psycopg2-binary>=2.9.0
gunicorn>=21.0.0
